\docType{data}
\name{HSMM}
\alias{HSMM}
\title{Single-cell expression data for HSMM cells}
\format{a \code{CellDataSet} object, as provided by the \code{monocle} package.}
\source{
monocle
}
\usage{
HSMM
}
\description{
This dataset contains expression profiles from a
time-series study of differentiating human skeletal muscle
myoblasts (HSMM). Expression values are in FPKM units, as
calculated with the Cufflinks package.
}
\author{
Cole Trapnell, 03-17-2014
}

