\name{HSMMSingleCell-package}
\alias{HSMMSingleCell-package}
\alias{HSMMSingleCell}
\docType{package}
\title{
Differential expression and time-series analysis for single-cell expression experiments.
}
\description{
Monocle is a set of tools for analyzing single-cell gene expression experiments. Monocle has been tested with RNA-Seq and qPCR, but could work with other types of data as well. The package provides basic differential expression and clustering tools, as well as more sophisticated tools. Monocle can place individual cells in the order of their progress through a biological process such as cell differentiation. This enables Monocle to find genes that are differentially expressed as a function of that process.
}
\details{
\tabular{ll}{
Package: \tab monocle\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-11-19\cr
License: \tab Boost\cr
}

}
\author{
Cole Trapnell
Maintainer: Cole Trapnell <cole@broadinstitute.org>
}
\references{
Trapnell, Cacchiarelli, et al
}
\keyword{ package }
\examples{
}